backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
pymongo>=4.6.0
dnspython>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0
imbalanced-learn>=0.11.0